Made changes to UI, Add clerk for authentication (WIP)
